useSearchParams() 사용 페이지에 export const dynamic = "force-dynamic"; 추가 #119
